requirements.txt
prettytable
python-nmap